Move _get_edge_qubits to PlaquetteQubits
